# clients/typescript/scripts/openapi.snapshot.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Spatial.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5299/"
    }
  ],
  "paths": {
...
    "/api/catalogue": {
      "get": {
        "tags": [
          "StoreEndpoints"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pattern",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets/{id}": {
      "get": {
        "tags": [
          "StoreEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDescription"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatasetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/features/scan": {
      "post": {
        "tags": [
          "StoreEndpoints"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureBatchesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/features/query": {
      "post": {
        "tags": [
          "StoreEndpoints"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureBatchesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/features/write": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureWriteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/begin": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/commit": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/rollback": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/sleep": {
      "post": {
        "tags": [
          "StoreEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BboxDto": {
        "required": [
          "minX",
          "minY",
          "maxX",
          "maxY"
        ],
        "type": "object",
        "properties": {
          "minX": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "minY": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxX": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxY": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "BeginTransactionResponse": {
        "required": [
          "transaction"
        ],
        "type": "object",
        "properties": {
          "transaction": {
            "type": "string"
          }
        }
      },
...
      "CatalogueResponse": {
        "required": [
          "datasets"
        ],
        "type": "object",
        "properties": {
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatasetSummary"
            }
          }
        }
      },
      "CreateDatasetRequest": {
        "required": [
          "dataset",
          "batch",
          "srid"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          },
          "batch": {
            "type": "string"
          },
          "srid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CreateDatasetResponse": {
        "required": [
          "dataset"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          }
        }
      },
...
      "DatasetDescription": {
        "required": [
          "id",
          "schemaName",
          "table",
          "geometryColumn",
          "srid",
          "geometryType",
          "estimatedRowCount",
          "idColumns",
          "schema"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "schemaName": {
            "type": "string"
          },
          "table": {
            "type": "string"
          },
          "geometryColumn": {
            "type": "string"
          },
          "srid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "geometryType": {
            "type": "string"
          },
          "estimatedRowCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "idColumns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/FeatureSchema"
          }
        }
      },
      "DatasetSummary": {
        "required": [
          "id",
          "schema",
          "table",
          "geometryColumn",
          "srid",
          "estimatedRowCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "table": {
            "type": "string"
          },
          "geometryColumn": {
            "type": "string"
          },
          "srid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "estimatedRowCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FeatureBatchesResponse": {
        "required": [
          "batches"
        ],
        "type": "object",
        "properties": {
          "batches": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FeatureQueryRequest": {
        "required": [
          "dataset"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          },
          "bbox": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BboxDto"
              }
            ]
          },
          "filter": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FeatureSchema": { },
      "FeatureWriteRequest": {
        "required": [
          "dataset",
          "batch"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          },
          "batch": {
            "type": "string"
          },
          "transaction": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FeatureWriteResponse": {
        "required": [
          "appended"
        ],
        "type": "object",
        "properties": {
          "appended": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ScanRequest": {
        "required": [
          "dataset"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          }
        }
      },
...
      "SleepRequest": {
        "required": [
          "milliseconds"
        ],
        "type": "object",
        "properties": {
          "milliseconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "SleepResponse": {
        "required": [
          "slept"
        ],
        "type": "object",
        "properties": {
          "slept": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "TransactionRequest": {
        "required": [
          "transaction"
        ],
        "type": "object",
        "properties": {
          "transaction": {
            "type": "string"
          }
        }
      },
      "TransactionResponse": {
        "required": [
          "ok"
        ],
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}